Revert "Issue #2498 Removed static lock from buffermanager impl"
