+++ dev/com.ibm.ws.logging.core/src/com/ibm/ws/collector/manager/buffer/BufferManagerImpl.java
@@ -12,6 +12,7 @@ package com.ibm.ws.collector.manager.buffer;
 
 import java.util.ArrayList;
 import java.util.HashSet;
+import java.util.List;
 import java.util.Queue;
 import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
@@ -27,7 +28,7 @@ public class BufferManagerImpl extends BufferManager {
 
     private static final TraceComponent tc = Tr.register(BufferManagerImpl.class);	
     private Buffer<Object> ringBuffer;
-    private final ReentrantReadWriteLock RERWLOCK = new ReentrantReadWriteLock(true);
+    private static final ReentrantReadWriteLock RERWLOCK = new ReentrantReadWriteLock(true);
     private Set<SynchronousHandler> synchronousHandlerSet = new HashSet<SynchronousHandler>();
 
     private final int capacity;
@@ -35,20 +36,23 @@ public class BufferManagerImpl extends BufferManager {
     private final String sourceId;
     /* Map to keep track of the next event for a handler */
     private final ConcurrentHashMap<String, HandlerStats> handlerEventMap = new ConcurrentHashMap<String, HandlerStats>();
+    private static List<BufferManager> bufferManagerList= new ArrayList<BufferManager>();
     
-    protected Queue<Object> earlyMessageQueue;
+    
+    private Queue<Object> earlyMessageQueue;
+    private volatile static boolean EMQRemovedFlag = false;
     private static final int EARLY_MESSAGE_QUEUE_SIZE=400;
-    
+    private static final int EMQ_TIMER = 60 * 5 * 1000; //5 Minute timer
 
     public BufferManagerImpl(int capacity, String sourceId) {
         super();
         RERWLOCK.writeLock().lock();
         try {
-            BufferManagerEMQHelper.addBufferManagerList(this);
+            bufferManagerList.add(this);
             ringBuffer=null;
             this.sourceId = sourceId;
             this.capacity = capacity;
-            if(!BufferManagerEMQHelper.getEMQRemovedFlag())
+            if(!BufferManagerImpl.EMQRemovedFlag)
                 earlyMessageQueue = new SimpleRotatingSoftQueue<Object>(new Object[EARLY_MESSAGE_QUEUE_SIZE]);
         }finally {
             RERWLOCK.writeLock().unlock();
@@ -220,7 +224,34 @@ public class BufferManagerImpl extends BufferManager {
             RERWLOCK.writeLock().unlock();
         }
     }
-   
+
+    public static void removeEMQTrigger(){
+        RERWLOCK.writeLock().lock();
+        try {
+            EMQRemovedFlag=true;
+            for(BufferManager i: bufferManagerList) {
+                ((BufferManagerImpl) i).removeEMQ();
+            }
+        }finally {
+            RERWLOCK.writeLock().unlock();
+        }
+    }
+
+    public static void removeEMQByTimer(){
+        new java.util.Timer().schedule(
+                new java.util.TimerTask() {
+                    @Override
+                    public void run() {
+                        BufferManagerImpl.removeEMQTrigger();
+                    }
+                },
+                BufferManagerImpl.EMQ_TIMER);
+    }
+    
+    public static boolean getEMQRemovedFlag() {
+        return EMQRemovedFlag;
+    }
+
     public static class HandlerStats {
 
         private final String handlerId;
