+++ dev/com.ibm.ws.logging/src/com/ibm/ws/logging/source/LogSource.java
@@ -21,7 +21,7 @@ import com.ibm.websphere.ras.DataFormatHelper;
 import com.ibm.websphere.ras.Tr;
 import com.ibm.websphere.ras.TraceComponent;
 import com.ibm.websphere.ras.annotation.Trivial;
-import com.ibm.ws.collector.manager.buffer.BufferManagerEMQHelper;
+import com.ibm.ws.collector.manager.buffer.BufferManagerImpl;
 import com.ibm.ws.logging.RoutedMessage;
 import com.ibm.ws.logging.WsLogHandler;
 import com.ibm.ws.logging.data.GenericData;
@@ -114,8 +114,8 @@ public class LogSource implements Source, WsLogHandler {
             LogRecord logRecord = routedMessage.getLogRecord();
             if (logRecord != null && bufferMgr != null) {
                 GenericData parsedMessage = parse(routedMessage);
-                if (!BufferManagerEMQHelper.getEMQRemovedFlag() && extractMessage(routedMessage, logRecord).startsWith("CWWKF0011I")) {
-                    BufferManagerEMQHelper.removeEMQTrigger();
+                if (!BufferManagerImpl.getEMQRemovedFlag() && extractMessage(routedMessage, logRecord).startsWith("CWWKF0011I")) {
+                    BufferManagerImpl.removeEMQTrigger();
                 }
                 bufferMgr.add(parsedMessage);
             }
